feat: add auto open perview
